# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hotel API",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/rooms/{id}": {
            "get": {
                "operationId": "id_room",
                "tags": [
                    "Rooms"
                ],
                "description": "Récupère une chambre par son id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_room": {
                                    "type": "integer"
                                },
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "operationId": "updateRoom",
                "tags": [
                    "Rooms"
                ],
                "description": "Met à jour une chambre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "room",
                        "in": "body",
                        "description": "Chambre à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_room": {
                                    "type": "integer"
                                },
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deleteRoom",
                "tags": [
                    "Rooms"
                ],
                "description": "Supprime une chambre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/payments/{id}" : {
            "get": {
                "operationId": "payments_id",
                "tags": [
                    "Payments"
                ],
                "description": "Récupère un paiement par son id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_payment": {
                                    "type": "integer"
                                },
                                "id_reservation": {
                                    "type": "integer"
                                },
                                "payment_date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "payment_amount": {
                                    "type": "number",
                                    "format": "float"
                                }
                            
                            
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                
                
                }
            },
            "patch" : {
                "operationId": "updatePayment",
                "tags": [
                    "Payments"
                ],
                "description": "Met à jour un paiement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "payment",
                        "in": "body",
                        "description": "Paiement à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_reservation": {
                                    "type": "integer"
                                },
                                "payment_date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "payment_amount": {
                                    "type": "number",
                                    "format": "float"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_payment": {
                                    "type": "integer"
                                },
                                "id_reservation": {
                                    "type": "integer"
                                },
                                "payment_date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "payment_amount": {
                                    "type": "number",
                                    "format": "float"
                                }
                            
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deletePayment",
                "tags": [
                    "Payments"
                ],
                "description": "Supprime un paiement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
               
        },
...
        "/services": {
            "get": {
                "operationId": "services",
                "tags": [
                    "Services"
                ],
                "description": "Récupère la liste des services",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_service": {
                                        "type": "integer"
                                    },
                                    "service_name": {
                                        "type": "string"
                                    },
                                    "service_price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post" : {
                "operationId": "addService",
                "tags": [
                    "Services"
                ],
                "description": "Ajoute un service",
                "parameters": [
                    {
                        "name": "service",
                        "in": "body",
                        "description": "Service à ajouter",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "service_name": {
                                    "type": "string"
                                },
                                "service_price": {
                                    "type": "number",
                                    "format": "float"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_service": {
                                    "type": "integer"
                                },
                                "service_name": {
                                    "type": "string"
                                },
                                "service_price": {
                                    "type": "number",
                                    "format": "float"
                                }
                            }
                            
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
        },
...
    }
}